Add missing problem description to source code.
The problem "git.java" doesn't feature the problem description anywere in the codebase
it only features in the original pull request. This commit puts the missing description
as a comment in the top of the class.

I've also removed the boilerplate "insert your licence in your settings" thing that was present.